+++ polymorphia-frontend/components/course/event-section/RewardsInfo.tsx
@@ -1,0 +1,3 @@
+// export default function RewardsInfo() {
+//   return <></>;
+// }
