+++ polymorphia-frontend/components/navigation/Sidebar.tsx
@@ -14,6 +14,9 @@ import clsx from 'clsx';
 import { animateSidebar } from '@/animations/Navigation';
 import { useQuery } from '@tanstack/react-query';
 import { EventSectionService } from '@/services/course/EventSectionService';
+import { EventSectionCore } from '@/interfaces/course/EventSectionInterfaces';
+import { MenuOption } from '@/interfaces/navigation/NavigationInterfaces';
+import { updateMenuItems } from '@/services/course/EventSectionUtils';
 
 export default function Sidebar() {
   const { isSidebarExpanded, setIsSidebarExpanded, isSidebarLocked } =
@@ -33,18 +36,7 @@ export default function Sidebar() {
 
   let menuItems = [...MainMenuItems];
   if (isSuccess) {
-    const courseItem = menuItems.filter(
-      (menuOption) => menuOption.text === 'Kurs'
-    )[0];
-    courseItem.link = `course/${eventSections[0].id}`;
-
-    courseItem.subItems = eventSections.map((eventSection) => {
-      // TODO: use correct courseID
-      return {
-        text: eventSection.name,
-        link: `course/${eventSection.id}`,
-      };
-    });
+    updateMenuItems(menuItems, eventSections);
   }
 
   return (
